Inline object instance behaviour
