+++ lib/toy/behaviours.rb
@@ -1,33 +1,5 @@
 module Toy
   module Behaviours
-    module InstanceVariables
-      def instance_variable_get(name)
-        ivar_map[name.to_sym]
-      end
-
-      def instance_variable_set(name, value)
-        ivar_map[name.to_sym] = value
-      end
-
-      def instance_variables
-        ivar_map.keys
-      end
-
-      def instance_variable_defined?(name)
-        ivar_map.has_key?(name.to_sym)
-      end
-
-      def remove_instance_variable(name)
-        ivar_map.delete(name.to_sym)
-      end
-
-      private
-
-      def ivar_map
-        @ivar_map ||= {}
-      end
-    end
-
     module Constants
       def const_get(name)
         name = name.to_sym
@@ -70,27 +42,6 @@ module Toy
 
       def method_map
         @method_map ||= {}
-      end
-    end
-
-    module ClassRelationships
-      def kind_of?(klass)
-        superclass = self.class
-        while superclass
-          return true if klass == superclass
-          superclass = superclass.superclass
-        end
-        false
-      end
-    end
-
-    module Inspection
-      def to_s
-        inspect
-      end
-
-      def inspect
-        "#<#{self.class}>"
       end
     end
   end
